Update to support Django 1.9
Django 1.9 removes ReverseManyRelatedObjectsDescriptor in favor of using ManyRelatedObjectsDescriptor with `reverse` argument passed into it (though the meaning of reverse has been reversed, hah).

It also adds a set() method to RelatedManager that handles setting all the new related objects. It tries to be more efficient by only adding and removing the ones necessary instead of clearing them all and adding them again. The old behavior is maintained by passing the `clear` keyword argument, which is what we now do.